+++ libs/@guardian/source-react-components/scripts/create-icons/index.ts
@@ -74,7 +74,6 @@ void (async () => {
 
 		await writeFile(
 			filepath,
-			// @ts-expect-error -- Some prettier config types are not narrowed.
 			await format(warning + component, { filepath, ...prettierConfig }),
 			'utf8',
 		);