+++ libs/@guardian/source-react-components/src/accordion/AccordionRowNoJS.tsx
@@ -3,6 +3,7 @@ import type { EmotionJSX } from '@emotion/react/types/jsx-namespace';
 import { visuallyHidden as _visuallyHidden } from '@guardian/source-foundations';
 import { SvgChevronDownSingle } from '../../vendor/icons/SvgChevronDownSingle';
 import type { Theme } from '../@types/Theme';
+import { mergeThemes } from '../utils/themes';
 import type { AccordionRowProps } from './AccordionRow';
 import {
 	accordionRow,
@@ -16,6 +17,11 @@ import {
 	toggleIconWithLabel,
 	toggleLabel,
 } from './styles';
+import type { ThemeAccordion } from './theme';
+import {
+	themeAccordion as defaultTheme,
+	transformProviderTheme,
+} from './theme';
 
 const visuallyHidden = css`
 	${_visuallyHidden}
@@ -32,17 +38,36 @@ export const AccordionRowNoJS = ({
 	hideToggleLabel = false,
 	children,
 	cssOverrides,
+	theme,
 	...props
 }: Omit<AccordionRowProps, 'onClick'>): EmotionJSX.Element => {
+	const mergedTheme = (providerTheme: Theme['accordion']) =>
+		mergeThemes<ThemeAccordion, Theme['accordion']>(
+			defaultTheme,
+			theme,
+			providerTheme,
+			transformProviderTheme,
+		);
 	return (
 		<div
-			css={(theme: Theme) => [accordionRow(theme.accordion), cssOverrides]}
+			css={(providerTheme: Theme) => [
+				accordionRow(mergedTheme(providerTheme.accordion)),
+				cssOverrides,
+			]}
 			{...props}
 		>
 			<label>
-				<input type="checkbox" css={noJsInput} role="button" />
+				<input
+					type="checkbox"
+					css={(providerTheme: Theme) =>
+						noJsInput(mergedTheme(providerTheme.accordion))
+					}
+					role="button"
+				/>
 				<div
-					css={(theme: Theme) => noJsButton(theme.accordion)}
+					css={(providerTheme: Theme) =>
+						noJsButton(mergedTheme(providerTheme.accordion))
+					}
 					data-target="label"
 				>
 					<strong css={labelText}>{label}</strong>
@@ -55,10 +80,15 @@ export const AccordionRowNoJS = ({
 							]}
 							data-target="toggle-label-show"
 						>
-							<span css={[toggleLabel, hideToggleLabel ? visuallyHidden : '']}>
+							<span
+								css={(providerTheme: Theme) => [
+									toggleLabel(mergedTheme(providerTheme.accordion)),
+									hideToggleLabel ? visuallyHidden : '',
+								]}
+							>
 								Show<span css={visuallyHidden}> more</span>
 							</span>
-							<SvgChevronDownSingle />
+							<SvgChevronDownSingle theme={{ fill: theme?.iconFill }} />
 						</div>
 						<div
 							css={[
@@ -68,10 +98,15 @@ export const AccordionRowNoJS = ({
 							]}
 							data-target="toggle-label-hide"
 						>
-							<span css={[toggleLabel, hideToggleLabel ? visuallyHidden : '']}>
+							<span
+								css={(providerTheme: Theme) => [
+									toggleLabel(mergedTheme(providerTheme.accordion)),
+									hideToggleLabel ? visuallyHidden : '',
+								]}
+							>
 								Hide
 							</span>
-							<SvgChevronDownSingle />
+							<SvgChevronDownSingle theme={{ fill: theme?.iconFill }} />
 						</div>
 					</div>
 				</div>