+++ libs/@guardian/source-react-components/src/button/theme.ts
@@ -1,5 +1,6 @@
 import { palette } from '@guardian/source-foundations';
 
+/** @deprecated Use Partial<ThemeButton> instead */
 export type ButtonTheme = {
 	textPrimary: string;
 	backgroundPrimary: string;
@@ -13,6 +14,20 @@ export type ButtonTheme = {
 	textSubdued?: string;
 };
 
+export type ThemeButton = {
+	textPrimary: string;
+	backgroundPrimary: string;
+	backgroundPrimaryHover: string;
+	textSecondary: string;
+	backgroundSecondary: string;
+	backgroundSecondaryHover: string;
+	textTertiary: string;
+	backgroundTertiaryHover: string;
+	borderTertiary: string;
+	textSubdued: string;
+};
+
+/** @deprecated Use `themeButton` and component `theme` prop instead of emotion's `ThemeProvider` */
 export const buttonThemeDefault: { button: ButtonTheme } = {
 	button: {
 		textPrimary: palette.neutral[100],
@@ -28,6 +43,7 @@ export const buttonThemeDefault: { button: ButtonTheme } = {
 	},
 };
 
+/** @deprecated Use `themeButtonBrand` and component `theme` prop instead of emotion's `ThemeProvider` */
 export const buttonThemeBrand: { button: ButtonTheme } = {
 	button: {
 		textPrimary: palette.brand[400],
@@ -43,6 +59,7 @@ export const buttonThemeBrand: { button: ButtonTheme } = {
 	},
 };
 
+/** @deprecated Use `themeButtonBrandAlt` and component `theme` prop instead of emotion's `ThemeProvider` */
 export const buttonThemeBrandAlt: { button: ButtonTheme } = {
 	button: {
 		textPrimary: palette.neutral[100],
@@ -51,8 +68,48 @@ export const buttonThemeBrandAlt: { button: ButtonTheme } = {
 		textSecondary: palette.neutral[7],
 		backgroundSecondary: palette.brandAlt[200],
 		backgroundSecondaryHover: '#F2AE00', // One-off colour variant generated from palette.brandAlt[200]
+		textTertiary: palette.neutral[0],
 		backgroundTertiaryHover: '#FFD213', // One-off colour variant
 		borderTertiary: palette.neutral[7],
 		textSubdued: palette.neutral[7],
 	},
 };
+
+export const themeButton: ThemeButton = {
+	textPrimary: palette.neutral[100],
+	backgroundPrimary: palette.brand[400],
+	backgroundPrimaryHover: '#234B8A', // One-off colour variant generated from palette.brand[400]
+	textSecondary: palette.brand[400],
+	backgroundSecondary: palette.brand[800],
+	backgroundSecondaryHover: '#ACC9F7', // One-off colour variant generated from palette.brand[800]
+	textTertiary: palette.brand[400],
+	backgroundTertiaryHover: '#E5E5E5', // One-off colour variant
+	borderTertiary: palette.brand[400],
+	textSubdued: palette.brand[400],
+} as const;
+
+export const themeButtonBrand: ThemeButton = {
+	textPrimary: palette.brand[400],
+	backgroundPrimary: palette.neutral[100],
+	backgroundPrimaryHover: '#E0E0E0', // One-off colour variant generated from palette.neutral[100]
+	textSecondary: palette.neutral[100],
+	backgroundSecondary: palette.brand[600],
+	backgroundSecondaryHover: '#234B8A', // One-off colour variant generated from palette.brand[600]
+	textTertiary: palette.neutral[100],
+	backgroundTertiaryHover: palette.brand[300],
+	borderTertiary: palette.neutral[100],
+	textSubdued: palette.neutral[100],
+} as const;
+
+export const themeButtonBrandAlt: ThemeButton = {
+	textPrimary: palette.neutral[100],
+	backgroundPrimary: palette.neutral[7],
+	backgroundPrimaryHover: '#454545', // One-off colour variant generated from palette.neutral[7]
+	textSecondary: palette.neutral[7],
+	backgroundSecondary: palette.brandAlt[200],
+	backgroundSecondaryHover: '#F2AE00', // One-off colour variant generated from palette.brandAlt[200]
+	textTertiary: palette.neutral[0],
+	backgroundTertiaryHover: '#FFD213', // One-off colour variant
+	borderTertiary: palette.neutral[7],
+	textSubdued: palette.neutral[7],
+} as const;