+++ libs/@guardian/source-react-components/src/link/ButtonLink.tsx
@@ -24,6 +24,7 @@ export const ButtonLink = ({
 	iconSide = 'left',
 	cssOverrides,
 	children,
+	theme,
 	...props
 }: ButtonLinkProps): EmotionJSX.Element => {
 	return (
@@ -34,6 +35,7 @@ export const ButtonLink = ({
 				iconSvg,
 				iconSide,
 				cssOverrides,
+				theme,
 			})}
 			{...props}
 		>